+++ test/test_common.py
@@ -181,3 +181,38 @@ class TestFishCommon(object):
         message = 'Hello HMAC'
         secret = '12345678'
         assert hmac_sha256(secret, message) == '5eb8bdabdaa43f61fb220473028e49d40728444b4322f3093decd9a356afd18f'
+
+    # test Base64() tc
+    def test_base64_01(self):
+        assert Base64.string('hello world') == b'aGVsbG8gd29ybGQ='
+    
+        here = os.path.split(os.path.dirname(__file__))[0]
+        file_path = os.path.join(here, 'test', '__init__.py')
+    
+        file_base64 = (b'aW1wb3J0IHN5cw0Kc3lzLnBhdGguYXBwZW5kKCcuLi9maXNoYmFzZScpDQoNCmZyb20gdGVzdC50'
+                       b'ZXN0X2ZpbGUgaW1wb3J0ICoNCmZyb20gdGVzdC50ZXN0X2NvbW1vbiBpbXBvcnQgKg0KZnJvbSB0'
+                       b'ZXN0LnRlc3RfZGF0ZSBpbXBvcnQgKg0KZnJvbSB0ZXN0LnRlc3Rfc3lzdGVtIGltcG9ydCAqDQoN'
+                       b'CnB5dGVzdC5tYWluKCkNCg==')
+    
+        assert Base64.file(file_path) == file_base64
+    
+        assert Base64.decode(b'aGVsbG8gd29ybGQ=') == b'hello world'
+
+    # test Base64()  tc
+    def test_base64_02(self):
+    
+        assert GetMD5.string('hello world') != b'aGVsbG8gd29ybGQ=='
+        assert Base64.decode(b'aGVsbG8gd29ybGQ=') != b'hello'
+    
+        here = os.path.split(os.path.dirname(__file__))[0]
+        file_path = os.path.join(here, 'test', '__init__.py')
+        file_path1 = os.path.join(here, 'test', '__init1__.py')
+    
+        if sys.version > '3':
+            with pytest.raises(FileNotFoundError):
+                GetMD5.file(file_path1)
+        else:
+            with pytest.raises(IOError):
+                GetMD5.file(file_path1)
+    
+        assert GetMD5.file(file_path) != b'bb7528c9778b2377e30b0f7e4c26fef0'