+++ docs/change_log.rst
@@ -7,6 +7,7 @@
 * 19046, setup, edit setup.py to add long description etc., the package detail;
 * issue ID use directly on github
 * 36, common, edit function :meth:`fish_common.check_str()`, optimize, doc and unittest;
+* 38, common, edit function :meth:`fish_common.find_files()`, optimize, doc and unittest;
 
 2018.6.6 v1.0.13
 ---------------------------