+++ docs/change_log.rst
@@ -11,6 +11,7 @@
 * 37, date, edit function :meth:`fish_date.get_years()`, optimize, doc and unittest;
 * 27, common, edit function :meth:`fish_common.hmac_sha256()`, optimize, doc and unittest;
 * 61, date, edit function :meth:`fish_date.get_date_range()`, optimize, doc and unittest;
+* 57, common, update function :meth:`fish_common.GetMD5.string()`, optimize;
 
 2018.6.6 v1.0.13
 ---------------------------